Add hardlink(1) runs to deduplicate backup directories.
I tend to make copies of directories from time to time and that gives
me duplicate copies.  I don't want those on the backup tapes, so run
hardlink(1) across the backup directories to reduce the space used.
Just running it now the backup directories went from 580G to 339G.  So
this is not insignificant.